requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
groq
pypdf
pandas
openpyxl
python-dotenv
